+++ _define.php
@@ -1,25 +1,26 @@
 <?php
-# -- BEGIN LICENSE BLOCK ----------------------------------
-#
-# This file is part of Related, a plugin for DotClear2.
-#
-# Licensed under the GPL version 2.0 license.
-# See LICENSE file or
-# http://www.gnu.org/licenses/old-licenses/gpl-2.0.html
-#
-# -- END LICENSE BLOCK ------------------------------------
-
-if (!defined('DC_RC_PATH')) return;
+/*
+ *  -- BEGIN LICENSE BLOCK ----------------------------------
+ *
+ *  This file is part of Related, a plugin for DotClear2.
+ *
+ *  Licensed under the GPL version 2.0 license.
+ *  See LICENSE file or
+ *  http://www.gnu.org/licenses/old-licenses/gpl-2.0.html
+ *
+ *  -- END LICENSE BLOCK ------------------------------------
+ */
 
 $this->registerModule(
-	/* Name */		"Related Pages",
-	/* Description*/	"Serve pages & scripts",
-	/* Author */		"Pep, contributors, Nicolas Roudaire",
-	/* Version */		'1.7.4',
-    /* Properties */	array('contentadmin,pages',
-                              'type' => 'plugin',
-							  'dc_min' => '2.7',
-							  'support' => 'http://forum.dotclear.net/viewtopic.php?id=48205',
-							  'details' => 'http://plugins.dotaddict.org/dc2/details/related'
-    )
+    "Related Pages", // Name
+    "Serve pages & scripts", // Description
+    "Pep, contributors, Nicolas Roudaire", // Author
+    '1.8.0', // Version
+    [
+        'permissions' => dcCore::app()->auth->makePermissions([dcAuth::PERMISSION_CONTENT_ADMIN, dcPages::PERMISSION_PAGES]),
+        'type' => 'plugin',
+        'dc_min' => '2.24',
+        'support' => 'http://forum.dotclear.net/viewtopic.php?id=48205',
+        'details' => 'http://plugins.dotaddict.org/dc2/details/related'
+    ]
 );