+++ _widgets.php
@@ -1,15 +1,15 @@
 <?php
-# -- BEGIN LICENSE BLOCK ----------------------------------
-#
-# This file is part of Related, a plugin for DotClear2.
-#
-# Licensed under the GPL version 2.0 license.
-# See LICENSE file or
-# http://www.gnu.org/licenses/old-licenses/gpl-2.0.html
-#
-# -- END LICENSE BLOCK ------------------------------------
+/*
+ *  -- BEGIN LICENSE BLOCK ----------------------------------
+ *
+ *  This file is part of Related, a plugin for DotClear2.
+ *
+ *  Licensed under the GPL version 2.0 license.
+ *  See LICENSE file or
+ *  http://www.gnu.org/licenses/old-licenses/gpl-2.0.html
+ *
+ *  -- END LICENSE BLOCK ------------------------------------
+ */
 
-if (!defined('DC_RC_PATH')) { return; }
-
-$core->addBehavior('initWidgets', array('widgetsRelated', 'init'));
-$core->addBehavior('initDefaultWidgets', array('widgetsRelated', 'initDefaultWidgets'));
+dcCore::app()->addBehavior('initWidgets', ['widgetsRelated', 'init']);
+dcCore::app()->addBehavior('initDefaultWidgets', ['widgetsRelated', 'initDefaultWidgets']);